+++ requirements/serializer.py
@@ -4,6 +4,7 @@ from plans.serializer import PlanSerializer
 from products.serializers import RateSerializer
 from products.models import Rate
 from customers.serializer import Customer_typeSerializer
+from customers.models import Customer_type
 from rest_framework import serializers
 class RequirementSerializer(serializers.ModelSerializer):
     class Meta:
@@ -34,10 +35,18 @@ class RatesRequirementSerializer(serializers.ModelSerializer):
     class Meta:
         model = RateRequirement
         fields = '__all__'
-        
+
 class RateWithRelatedDataSerializer(serializers.ModelSerializer):
-    rate_requirements = RateRequirementSerializer(many=True, read_only=True)
+    customer_type = serializers.SerializerMethodField()
+    requirements = serializers.SerializerMethodField()
 
     class Meta:
         model = Rate
         fields = '__all__'
+
+    def get_customer_type(self, obj):
+        return Customer_type.objects.filter(raterequirement__rate=obj, raterequirement__is_active=True).distinct().values_list('name', flat=True)
+
+    def get_requirements(self, obj):
+        return Requirement.objects.filter(raterequirement__rate=obj, raterequirement__is_active=True).distinct().values_list('requirement_name', flat=True)
+
