+++ requirements/urls.py
@@ -1,6 +1,6 @@
 from django.urls import include, path
 from rest_framework import routers
-from .views import RateRequirement_Api, RateRequirement_Detail, All_Rates, RateWithRelatedDataView
+from .views import RateRequirement_Api, RateRequirement_Detail, RateWithRelatedDataView
 urlpatterns = [
     path('rates/', RateRequirement_Api.as_view()),
     path('rates/<str:pk>', RateRequirement_Detail.as_view()),
